refactor(converters): extract `$defs` alias resolution into `_refs.py`
